consolidate App and ActualApp
